Add unique fd gen

+++ mirrord-protocol/src/codec.rs
@@ -39,7 +39,7 @@ pub struct LogMessage {
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct ReadFileRequest {
-    pub fd: i32,
+    pub fd: usize,
     pub buffer_size: usize,
 }
 
@@ -112,26 +112,26 @@ pub struct OpenFileRequest {
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct OpenRelativeFileRequest {
-    pub relative_fd: i32,
+    pub relative_fd: usize,
     pub path: PathBuf,
     pub open_options: OpenOptionsInternal,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct SeekFileRequest {
-    pub fd: i32,
+    pub fd: usize,
     pub seek_from: SeekFromInternal,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct WriteFileRequest {
-    pub fd: i32,
+    pub fd: usize,
     pub write_bytes: Vec<u8>,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct CloseFileRequest {
-    pub fd: i32,
+    pub fd: usize,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
@@ -156,7 +156,7 @@ pub enum ClientMessage {
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct OpenFileResponse {
-    pub fd: i32,
+    pub fd: usize,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
